requirements-backend.txt
fastapi
httpx
matplotlib
numpy
pandas
python-multipart
PyYAML
torch
uvicorn
